[TIDY] Refactor renderlib into two parts
Most of the make scripts only need the schedule part of the renderlib.
The schedule part has way less dependencies (e.g. no ImageMagick), so
those scripts can be used without them now.
